devices_backend/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
contextlib2==21.6.0
Django==5.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-ninja==1.1.0
django-ninja-extra==0.20.3
injector==0.21.0
pydantic==2.5.3
pydantic_core==2.14.6
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
